requirements.txt
boto3>=1.34.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
APScheduler>=3.10.4
